+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fixed band, result has same number of bits as its first argument
 - fixed bor, result has same number of bits as its first argument
 - fixed bxor, result has same number of bits as its first argument
+- fixed contrast, 1.2.3.0/24 is really to the left of 1.2.4.0/24
 
 # 0.1.0
 - initial version